+++ demos/nm_students/index.php
@@ -1,0 +1,16 @@
+<?php
+
+require_once 'include/theme.php';
+
+
+
+include __DIR__ . '/../includes/header.php';
+?>
+
+<!-- Deine Demo-HTML hier -->
+<form action="create_data.php" method="get">
+  <button type="submit" class="btn btn-outline-primary">📊 Daten erstellen</button>
+</form>
+
+
+<?php include __DIR__ . '/../includes/footer.php'; ?>